fix(node_framework): Run gas adjuster task only if necessary (#2266)
## What ❔

Adds a check that would prevent gas adjuster task from actually running
if there are no users of `GasAdjuster`.

## Why ❔

`GasAdjuster` is provided as a resource for anyone to use, but its
support task uses RPC (which has usage limits), so it doesn't make sense
to run it if nobody uses `GasAdjuster`.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.