feat(node): Port (most of) Node to the Node Framework (#2196)
## What ❔

Ports most of the EN parts to the framework, except:
- Reorg detector
- Genesis / snapshot recovery
- Unique metrics stuff

## Why ❔

This forms the "main" body of the framework-based EN. The binary can
already run, it can proxy and re-execute transactions, etc.
The remaining work may be somewhat complex, so I would like to ship it
separately -- so that it receives thorough review.

There are some TODOs without a task number in this PR, expect them
either to be fixed in the next one, or otherwise an issue to be added.

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zk fmt` and `zk lint`.
- [ ] Spellcheck has been run via `zk spellcheck`.