chore: use simple StorageLog struct instead of LogQuery when possible (#2252)
There are many fields in LogQuery on the core side (it also exists in
zk_evm) that are unused and often just set to some garbage value but
that is very confusing.

This PR changes the uses of LogQuery in the VmInterface to StorageLog,
which only has the necessary fields.

---------

Co-authored-by: perekopskiy <mikeson.dp@gmail.com>