chore: Fix path resolution when running from compiled binary (#2788)
Follow-up to #2785

We use `zksync_utils::env::locate_workspace` in our code, which behaves
differently locally (uses `cargo locate-project` locally and does a
fallback to `.` in docker), which is then altered via `..` in
`core_workspace_dir_or_current_dir`, making the path resolution very
hard.

I fix the issue so that it works both locally and in docker, and also
add comments explaining how it works.