feat(proof-data-handler): add tee_proof_generation_timeout_in_secs param (#3128)
## What ❔

Add `tee_proof_generation_timeout_in_secs` parameter to the
`proof-data-handler` configuration to avoid sharing the same
`proof_generation_timeout_in_secs` timeout with the ZK prover.

This timeout is for retrying TEE proof generation if it fails. Retries
continue indefinitely until successful.

## Why ❔

The TEE prover is much faster than the ZK prover, so some of the ZK
timeouts are too long to be shared with the TEE-specific code.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.