feat(da-clients): add EigenDA client (#3155)
This PR adds an EigenDA client.

The implementation uses the gRPC streams to send the authenticated
requests to dispatch the blob.
The protogen situation is very similar to Celestia, we use the generated
files as a temporary solution until there is a separate crate that
provides those.

This kind of function can be used to generate them in the future:
```rust
pub fn compile_protos() {
    let fds = protox::compile(
        [
            "proto/common.proto",
            "proto/disperser.proto",
        ],
        ["."],
    )
    .expect("protox failed to build");

    tonic_build::configure()
        .build_client(true)
        .build_server(false)
        .skip_protoc_run()
        .out_dir("generated")
        .compile_fds(fds)
        .unwrap();
}
```

Example config:
```
da_client:
  eigen:
    rpc_node_url: https://disperser-holesky.eigenda.xyz:443
    inclusion_polling_interval_ms: 10000
```
secrets:
```
da:
  eigen:
    private_key: PRIVATE_KEY_WITHOUT_0x_PREFIX
```

To enable EigenDA in ZK stack

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [ ] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [ ] Tests for the changes have been added / updated.
- [ ] Documentation comments have been added / updated.
- [ ] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.