feat: vm2 account validation (#2863)
Implements an account validation gas limit and the validation tracer for
vm2, along with better tests for validation.

Instead of a second gas limit like in vm_latest, the normal gas limit is
used. Unfortunately this means that the VM is not safe to use in the
sequencer until we forbid the use of gasleft. I didn't do it here
because it requires something like taint analysis and could break
existing contracts that didn't know that gasleft is forbidden.

---------

Co-authored-by: Alex Ostrovski <aov@matterlabs.dev>
Co-authored-by: Alex Ostrovski <slowli@users.noreply.github.com>