feat: (DB migration) Rename recursion_scheduler_level_vk_hash to snark_wrapper_vk_hash (#2809)
We have a configuration field `recursion_scheduler_level_vk_hash` which
actually stores `snark_wrapper_vk_hash` inside.
It happened because an old config value was utilized for the new purpose
some time ago.
This PR changes the name of field in a non-breaking way:

- `serde` (de)serialization happens with both `alias` and
`rename(serialize = "..")`, so that we serialize the field the same way
as before, but can deserialize either way. This is used for env configs
and API.
- `protobuf` deserialization is done by introducing a new field, and
reading whatever one is available.
- `protobuf` serialization always produced the _new_ field, so newly
generated configs should have new field name.
- ~~⚠️ DB column names was left as-is, because renaming DB columns is
not a trivial process.~~
- Upd: Migration was added. It copies the old column to the new one and
switches to the new one right away.

Having incorrect name that doesn't represent the value stored is
confusing and can lead to errors.