feat(vm): Split old and new VM implementations (#2915) ## What ❔
Splits old and new VM implementations in the `multivm` crate:

- Old VMs are encapsulated in the `LegacyVmInstance` enum, while new
ones in the `FastVmInstance` enum (which includes plain and shadowed VM
variants).
- Fast VM and `FastVmInstance` now expose a tracer type.
- Usage of the Fast VM in the batch executor are updated
correspondingly.

## Why ❔

It seems infeasible to unify the tracer model for old and new VMs, so
keeping them all in a single enum makes little sense.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zk fmt` and `zk lint`.