fix(test_contracts): dirty fix the `svm-rs-builds` broken compilation (#314)
## What ❔

- Disable the raw contracts generated by broken `svm-rs-builds` in
pre-compile step.

## Why ❔

- The `svm-rs-build` library is no longer valid. It generates the
invalid `build.rs` file (which contains duplicated constants) in
pre-compile phase. This is a workaround just to unblock the compilation
of the `core` module since the `svm-rs-build` is primary used in load
tests. Currently there is no other way of fixing this because latest
version of `svm-rs-builds` is not supported with the current version of
rust compiler.
- This workaround removes the support in `loadnext` to deploy and
interact with contracts.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.