fix(reorg_detector): reorg detector check window of blocks not just the tail  (#323)
## What ❔

- Fix the reorg logic to verify a window of blocks and not just the tail
of the chain.

## Why ❔

Previously, the reorg detector validated only the last block hash by
comparing it against the BTC node. Relying solely on the chain tail can
lead to edge cases. For example, during an ongoing reorg, the tail may
appear valid at the time of the check, but become invalid while the
database update is in progress. This race condition can cause the
sequencer or verifier to miss the reorg entirely. This fix allows to
verify a window of blocks (last 25 blocks) which helps to detect the
reorgs.

## Checklist

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [X] Documentation comments have been added / updated.
- [X] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Co-authored-by: 0xatomFusion <179967211+0xatomFusion@users.noreply.github.com>