fix(via_btc_watch): via btc watch skip blocks after reorg (#332)
## What ❔

- Add a validation to the btc watch to return if the to_block is less
than the from_block. This can happens when a reorg block is behind the
btc_watch.
- Add a validation to update the affected l1 block based on the last
processed block by the btc_watch.

## Why ❔

Before this PR, during a reorg, the l1_block number was set to the last
block affected by the reorg. If the BTC watcher’s last processed block
was behind this block, the intermediate blocks were skipped, which
caused the BTC watcher to miss blocks during indexing.

## Checklist

- [X] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [X] Tests for the changes have been added / updated.
- [X] Documentation comments have been added / updated.
- [X] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Co-authored-by: 0xatomFusion <179967211+0xatomFusion@users.noreply.github.com>